Enables multiplatform support
