+++ compiler-integration-test/src/commonTest/kotlin/land/sungbin/composeinvestigator/compiler/test/awaitInvalidation.kt
@@ -1,0 +1,9 @@
+// Copyright 2024 Ji Sungbin
+// SPDX-License-Identifier: Apache-2.0
+package land.sungbin.composeinvestigator.compiler.test
+
+import kotlinx.coroutines.test.TestScope
+
+@Suppress("NOTHING_TO_INLINE")
+inline fun TestScope.awaitInvalidation() =
+  testScheduler.advanceUntilIdle()
