+++ gradle.properties
@@ -11,6 +11,8 @@ kotlin.code.style=official
 kotlin.build.report.output=json
 kotlin.build.report.json.directory=.kotlin
 
+kotlin.suppressGradlePluginWarnings=IncorrectCompileOnlyDependencyWarning
+
 # https://kotlinlang.org/docs/dokka-migration.html#set-the-opt-in-flag
 org.jetbrains.dokka.experimental.gradle.pluginMode=V2Enabled
 org.jetbrains.dokka.experimental.gradle.pluginMode.noWarn=true
