+++ gradle/libs.versions.toml
@@ -1,4 +1,6 @@
 [versions]
+jdk = "22"
+
 gradle-android = "8.7.0"
 gradle-publish-maven = "0.29.0"
 gradle-test-logging = "4.0.0"
@@ -7,13 +9,11 @@ kotlin-core = "2.1.0-Beta1"
 kotlin-dokka = "2.0.0-Beta"
 kotlin-coroutines = "1.9.0"
 
-jetbrains-annotation = "26.0.0"
-
 androidx-activity = "1.9.2"
+androidx-annotation = "1.8.2"
 
 compose-activity = "1.9.2"
-compose-stableMarker = "1.0.5"
-compose-runtime = "1.7.3"
+compose-multiplatform = "1.7.0-rc01"
 
 spotless = "6.25.0"
 ktlint = "1.3.1"
@@ -24,27 +24,24 @@ test-kluent = "1.73"
 
 [plugins]
 gradle-publish-maven = { id = "com.vanniktech.maven.publish", version.ref = "gradle-publish-maven" }
+gradle-android-application = { id = "com.android.application", version.ref = "gradle-android" }
 gradle-test-logging = { id = "com.adarshr.test-logger", version.ref = "gradle-test-logging" }
 
 kotlin-dokka = { id = "org.jetbrains.dokka", version.ref = "kotlin-dokka" }
-kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin-core" }
+
+compose-multiplatform = { id = "org.jetbrains.compose", version.ref = "compose-multiplatform" }
 
 spotless = { id = "com.diffplug.spotless", version.ref = "spotless" }
 
 [libraries]
-gradle-android = { module = "com.android.tools.build:gradle", version.ref = "gradle-android" }
-
-jetbrains-annotation = { module = "org.jetbrains:annotations", version.ref = "jetbrains-annotation" }
+kotlin-coroutines = { module = "org.jetbrains.kotlinx:kotlinx-coroutines-core", version.ref = "kotlin-coroutines" }
 
 androidx-activity = { module = "androidx.activity:activity-ktx", version.ref = "androidx-activity" }
+androidx-annotation = { module = "androidx.annotation:annotation", version.ref = "androidx-annotation" }
 
 compose-activity = { module = "androidx.activity:activity-compose", version.ref = "compose-activity" }
-
-# TODO Support for Compose Multiplatform
-#  https://github.com/jisungbin/ComposeInvestigator/issues/184
-compose-runtime = { module = "androidx.compose.runtime:runtime", version.ref = "compose-runtime" }
-compose-runtimeStubs = { module = "androidx.compose.runtime:runtime-jvmstubs", version.ref = "compose-runtime" }
-compose-material = { module = "androidx.compose.material:material", version.ref = "compose-runtime" }
+compose-runtime = { module = "org.jetbrains.compose.runtime:runtime", version.ref = "compose-multiplatform" }
+compose-material = { module = "org.jetbrains.compose.material:material", version.ref = "compose-multiplatform" }
 
 test-diffutil = { module = "io.github.java-diff-utils:java-diff-utils", version.ref = "test-diffutil" }
 test-assertk = { module = "com.willowtreeapps.assertk:assertk", version.ref = "test-assertk" }
