+++ runtime/build.gradle.kts
@@ -1,7 +1,10 @@
 // Copyright 2024 Ji Sungbin
 // SPDX-License-Identifier: Apache-2.0
+import org.gradle.kotlin.dsl.dokka
+import org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi
+
 plugins {
-  kotlin("jvm")
+  kotlin("multiplatform")
   alias(libs.plugins.kotlin.dokka)
   id(libs.plugins.gradle.publish.maven.get().pluginId)
 }
@@ -21,19 +24,43 @@ dokka {
   }
 }
 
+@OptIn(ExperimentalKotlinGradlePluginApi::class)
 kotlin {
   explicitApi()
+  jvmToolchain(libs.versions.jdk.get().toInt())
+
   compilerOptions {
     optIn.addAll(
       "land.sungbin.composeinvestigator.runtime.ComposeInvestigatorCompilerApi",
       "land.sungbin.composeinvestigator.runtime.ExperimentalComposeInvestigatorApi",
     )
   }
-}
 
-dependencies {
-  compileOnly(libs.compose.runtimeStubs)
+  jvm()
 
-  testImplementation(kotlin("test-junit5", version = libs.versions.kotlin.core.get()))
-  testImplementation(libs.test.assertk)
+  iosArm64()
+  iosX64()
+  iosSimulatorArm64()
+
+  sourceSets {
+    commonMain {
+      dependencies {
+        compileOnly(libs.compose.runtime)
+        implementation(libs.androidx.annotation)
+      }
+    }
+
+    commonTest {
+      dependencies {
+        implementation(kotlin("test", version = libs.versions.kotlin.core.get()))
+        implementation(libs.test.assertk)
+      }
+    }
+
+    jvmTest {
+      dependencies {
+        implementation(kotlin("reflect", version = libs.versions.kotlin.core.get())) // Used by assertk
+      }
+    }
+  }
 }
