+++ runtime/src/commonTest/kotlin/land/sungbin/composeinvestigator/runtime/StabilityTest.kt
