--- runtime/src/main/kotlin/land/sungbin/composeinvestigator/runtime/ComposableScope.kt
@@ -1,9 +1,0 @@
-// Copyright 2024 Ji Sungbin
-// SPDX-License-Identifier: Apache-2.0
-package land.sungbin.composeinvestigator.runtime
-
-/** APIs annotated with this should only be used within Composable functions. */
-@MustBeDocumented
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY, AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
-public annotation class ComposableScope
