+++ samples/android/src/main/AndroidManifest.xml
